EdgeSolution/modules/PythonSampleLogs/requirements.txt
azure-iot-device~=2.0.0
pytz